Move change_dropout_strength to model_utils
